chore(scheduler): Restore scheduler API / compat with cujojs/most scheduler
