+++ packages/types/index.d.ts
@@ -29,7 +29,11 @@ export type Offset = number
 
 export interface Scheduler {
   now(): Time;
-  schedule(offset: Offset, delay: Delay, period: Period, task: Task): ScheduledTask;
+  asap (task: Task): ScheduledTask;
+  delay (delay: Delay, task: Task): ScheduledTask;
+  periodic (period: Period, task: Task): ScheduledTask;
+  schedule (delay: Delay, period: Period, task: Task): ScheduledTask;
+  scheduleTask (offset: Offset, delay: Delay, period: Period, task: Task): ScheduledTask;
   relative(offset: Offset): Scheduler;
   cancel(task: ScheduledTask): void;
   cancelAll(predicate: (task: ScheduledTask) => boolean): void;
